test(gatsby-script): Migrate e2e test to production and develop runtime suites (#35663)
* test(gatsby-script): Production runtime

* Develop runtime
