+++ e2e-tests/development-runtime/gatsby-script-records.js
@@ -1,0 +1,12 @@
+/**
+ * Naming matches `PerformanceResourceTiming` interface.
+ * @see {@link https://developer.mozilla.org/en-US/docs/Web/API/PerformanceResourceTiming}
+ */
+export const resourceRecord = {
+  fetchStart: `fetch-start`,
+  responseEnd: `response-end`,
+}
+
+export const markRecord = {
+  executeStart: `execute-start`,
+}
