+++ e2e-tests/development-runtime/package.json
@@ -50,7 +50,7 @@
     "update:cms-webhook": "node scripts/cms-webhook.js",
     "update:preview": "curl -X POST -d \"{ \\\"fake-data-update\\\": true }\" -H \"Content-Type: application/json\" http://localhost:8000/__refresh",
     "start-server-and-test": "start-server-and-test develop http://localhost:8000 cy:run",
-    "start-server-and-debug": "start-server-and-test develop http://localhost:8000 cy:open",
+    "start-server-and-test:locally": "start-server-and-test develop http://localhost:8000 cy:open",
     "cy:open": "cypress open",
     "cy:run": "node ../../scripts/cypress-run-with-conditional-record-flag.js --browser chrome"
   },
