--- e2e-tests/gatsby-script/.gitignore
@@ -1,11 +1,0 @@
-# Project dependencies
-.cache
-node_modules
-yarn-error.log
-
-# Build directory
-/public
-.DS_Store
-
-# Cypress output
-cypress/videos/
