--- e2e-tests/gatsby-script/README.md
@@ -1,10 +1,0 @@
-# Gatsby Script e2e test
-
-This test only runs in build + serve mode since `develop` makes it challenging to assert script loading.
-
-To test locally:
-
-1. `npm install`
-2. `npm run test:locally`
-
-Then click on the suites you want to run in Cypress.
