--- e2e-tests/gatsby-script/cypress.json
@@ -1,7 +1,0 @@
-{
-  "baseUrl": "http://localhost:9000",
-  "env": {
-    "GATSBY_COMMAND": "build"
-  },
-  "defaultCommandTimeout": 10000
-}
