--- e2e-tests/gatsby-script/gatsby-browser.tsx
@@ -1,1 +1,0 @@
-// TODO - Test case where we add <Script> in callbacks here
