--- e2e-tests/gatsby-script/gatsby-config.ts
@@ -1,11 +1,0 @@
-import type { GatsbyConfig } from "gatsby"
-
-const config: GatsbyConfig = {
-  siteMetadata: {
-    title: `e2e-gatsby-script`,
-    siteUrl: `https://gatsbyjs.com`,
-  },
-  plugins: [],
-}
-
-export default config
