--- e2e-tests/gatsby-script/src/hooks/use-occupy-main-thread.ts
@@ -1,16 +1,0 @@
-import React, { useEffect } from "react"
-
-/**
- * Attempts to occupy the main thread with work so that the idle strategy can be observed.
- */
-export function useOccupyMainThread(timeout: number = 3000) {
-  useEffect(() => {
-    const interval = setInterval(() => {
-      console.log(`Occupying main thread`)
-    }, 0)
-
-    setTimeout(() => {
-      clearInterval(interval)
-    }, timeout)
-  }, [])
-}
