--- e2e-tests/gatsby-script/src/pages/404.tsx
@@ -1,15 +1,0 @@
-import * as React from "react"
-import { Link } from "gatsby"
-import "../styles/global.css"
-
-const NotFoundPage = () => {
-  return (
-    <main>
-      <title>404</title>
-      <h1>404</h1>
-      <Link to="/">Back to index</Link>
-    </main>
-  )
-}
-
-export default NotFoundPage
