--- e2e-tests/gatsby-script/src/styles/global.css
@@ -1,32 +1,0 @@
-:root {
-  --text: #000;
-  --background: #fff;
-  --line: #e8e3f1;
-}
-
-@media (prefers-color-scheme: dark) {
-  :root {
-    --text: #fff;
-    --background: #000;
-    --line: #424464;
-  }
-}
-
-body {
-  font-family: monospace;
-  font-size: 15px;
-  color: var(--text);
-  background-color: var(--background);
-}
-
-table,
-th,
-td {
-  border: 1px dashed var(--line);
-  border-collapse: collapse;
-}
-
-th,
-td {
-  padding: 0.5em;
-}
