--- e2e-tests/gatsby-script/src/utils/trim.ts
@@ -1,3 +1,0 @@
-export function trim(number: number): number {
-  return Math.round(Number(number))
-}
