+++ e2e-tests/production-runtime/package.json
@@ -6,18 +6,18 @@
   "dependencies": {
     "babel-plugin-search-and-replace": "^1.1.0",
     "cypress": "^6.5.0",
-    "gatsby": "^4.1.6",
-    "gatsby-cypress": "^2.1.0",
-    "gatsby-plugin-image": "^2.1.3",
-    "gatsby-plugin-less": "^6.1.0",
-    "gatsby-plugin-manifest": "^4.1.4",
-    "gatsby-plugin-offline": "^5.1.4",
-    "gatsby-plugin-react-helmet": "^5.1.0",
-    "gatsby-plugin-sass": "^5.1.1",
-    "gatsby-plugin-sharp": "^4.1.4",
-    "gatsby-plugin-stylus": "^4.1.0",
+    "gatsby": "next",
+    "gatsby-cypress": "next",
+    "gatsby-plugin-image": "next",
+    "gatsby-plugin-less": "next",
+    "gatsby-plugin-manifest": "next",
+    "gatsby-plugin-offline": "next",
+    "gatsby-plugin-react-helmet": "next",
+    "gatsby-plugin-sass": "next",
+    "gatsby-plugin-sharp": "next",
+    "gatsby-plugin-stylus": "next",
     "gatsby-seo": "^0.1.0",
-    "gatsby-source-filesystem": "^4.1.3",
+    "gatsby-source-filesystem": "next",
     "glob": "^7.1.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -41,6 +41,7 @@
     "test:offline": "npm run build:offline && yarn start-server-and-test:offline && npm run test-env-vars",
     "test-env-vars": " node __tests__/env-vars.js",
     "start-server-and-test": "start-server-and-test serve http://localhost:9000 cy:run",
+    "start-server-and-test:locally": "start-server-and-test serve http://localhost:9000 cy:open",
     "start-server-and-test:offline": "start-server-and-test serve http://localhost:9000 cy:run:offline",
     "cy:open": "cypress open",
     "cy:open:offline": "npm run cy:open -- --env TEST_PLUGIN_OFFLINE=y",
