+++ e2e-tests/production-runtime/src/utils/trim.js
@@ -1,0 +1,3 @@
+export function trim(number) {
+  return Math.round(Number(number))
+}
